Moved common directory to device and got rid of aie::common namespace
