+++ src/runtime_src/xdp/profile/plugin/aie_debug/CMakeLists.txt
@@ -13,8 +13,8 @@ file(GLOB AIE_DEBUG_PLUGIN_FILES
 )
 
 file(GLOB AIE_DRIVER_COMMON_UTIL_FILES
-  "${PROFILE_DIR}/plugin/common/*.h"
-  "${PROFILE_DIR}/plugin/common/*.cpp"
+  "${PROFILE_DIR}/device/common/*.h"
+  "${PROFILE_DIR}/device/common/*.cpp"
 )
 
 if (XDP_MINIMAL_BUILD STREQUAL "yes")
