+++ src/runtime_src/xdp/profile/plugin/aie_debug/aie_debug_plugin.h
@@ -7,7 +7,7 @@
 #include <boost/property_tree/ptree.hpp>
 #include <memory>
 
-#include "xdp/profile/plugin/common/client_transaction.h"
+#include "xdp/profile/device/common/client_transaction.h"
 #include "xdp/profile/database/static_info/aie_constructs.h"
 #include "xdp/profile/database/static_info/filetypes/base_filetype_impl.h"
 #include "xdp/profile/plugin/vp_base/vp_base_plugin.h"
@@ -45,7 +45,7 @@ namespace xdp {
       {module_type::mem_tile, "Memory Tile"}
     };
     
-    std::unique_ptr<aie::common::ClientTransaction> transactionHandler;
+    std::unique_ptr<aie::ClientTransaction> transactionHandler;
     uint8_t* txn_ptr;
     XAie_DevInst aieDevInst = {0};
     boost::property_tree::ptree aie_meta;
