+++ src/runtime_src/xdp/profile/plugin/aie_profile/CMakeLists.txt
@@ -30,8 +30,8 @@ file(GLOB AIE_PROFILE_UTIL_FILES
 )
 
 file(GLOB AIE_DRIVER_COMMON_UTIL_FILES
-  "${PROFILE_DIR}/plugin/common/*.h"
-  "${PROFILE_DIR}/plugin/common/*.cpp"
+  "${PROFILE_DIR}/device/common/*.h"
+  "${PROFILE_DIR}/device/common/*.cpp"
 )
 
 if (XDP_MINIMAL_BUILD STREQUAL "yes")
