--- src/runtime_src/xdp/profile/plugin/common/transactions/op_defs.h
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (C) 2023 Advanced Micro Devices, Inc. - All rights reserved
-#ifndef __OP_DEFS_H__
-#define __OP_DEFS_H__
-
-#define OP_LIST(OP) \
-        OP(TRANSACTION_OP) \
-        OP(WAIT_OP) \
-        OP(PENDINGBDCOUNT_OP) \
-        OP(DBGPRINT_OP) \
-        OP(PATCHBD_OP)
-
-#include "op_base.h"
-#endif
